pyproject.toml
[tool.poetry]
name = "python-tire-degradation"
version = "0.1.0"
description = ""
authors = ["Patrick Kalkman <patrick@simpletechture.nl>"]
license = "MIT"
readme = "README.md"
packages = [{include = "python_tire_degradation"}]

[tool.poetry.dependencies]
python = "^3.11"
fastf1 = "^3.0.3"
pandasai = "^0.2.14"
seaborn = "^0.12.2"
requests = "^2.31.0"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
